requirements.txt
flask
pycryptodome